Attempt to get rootless docker in LXC
This long-standing battle between rootless LXC with rootless Docker inside
resulted in nothing, in the end. The ending problem seemed to be that rootless
Docker needs write access to /etc/subuid; this is already being used by
the rootless LXC and is limited as such. The Docker systemd (user) service
cannot write to the path and fails to start as such.