Use rbind instead of bind to access ct mounted ZFS data
rbind is basically a recursive bind, which helps binding all subfolders
inside a ZFS dataset. Without rbind, you are not able to (volume) mount
a subdirectory from within a nested Docker container.

rbinds are unfortuantely not native to Proxmox but to LXC. This means
that the mount is not visible from the web GUI.