WIP: First attempt at restructuring Ansible
The previous layout was already quite messy and became to overview and
expand upon. This change will start a reorganization of the code, which
should enable a better way to integrate the tofu and Ansible parts
automatically.
This relies on all docker setups being the same, and overrides being
messy, esp. on the host side.